+++ climaApp/app/src/main/java/com/example/climachallenge/retrofit/models/OpenWeatherResponse.kt
@@ -9,5 +9,5 @@ data class OpenWeatherResponse(
     val lat: Double,
     val lon: Double,
     val timezone: String,
-    val timezone_offset: Int
+    val timezone_offset: Double
 )
